update command naming and unnest ifs
