+++ src/main/java/frc/robot/drivetrain/commands/XBoxDriveCommand.java
@@ -7,7 +7,7 @@ import edu.wpi.first.math.filter.SlewRateLimiter;
 import edu.wpi.first.wpilibj.XboxController;
 import frc.robot.drivetrain.Drivetrain;
 
-public class XBoxDrive extends Drive {
+public class XBoxDriveCommand extends DriveCommand {
 
   XboxController m_controller;
 
@@ -16,7 +16,7 @@ public class XBoxDrive extends Drive {
   private final SlewRateLimiter m_yspeedLimiter = new SlewRateLimiter(3);
   private final SlewRateLimiter m_rotLimiter = new SlewRateLimiter(3);
 
-  public XBoxDrive(Drivetrain subsystem, XboxController joysticks) {
+  public XBoxDriveCommand(Drivetrain subsystem, XboxController joysticks) {
     super(subsystem);
     this.m_controller = joysticks;
   }
